+++ src/app/_layout.tsx
@@ -7,26 +7,23 @@ import Toast from "react-native-toast-message";
 import { toastConfig } from "../components/feedback/ToastMessage";
 import AuthCheck from "../components/AuthCheck";
 import { BackgroundColorProvider } from "../contexts/BackgroundColorProvider";
-import { MoodKeywordProvider } from "../contexts/keywordProvider";
 
 export default function RootLayout() {
   return (
     <>
       {/* // ThemeProvider를 통해 앱 전반에 다크테마 / 라이트테마 적용 가능함 */}
       <ThemeProvider value={DefaultTheme}>
-        <MoodKeywordProvider>
-          <BackgroundColorProvider>
-            <StatusBar style="dark" />
-            {/* RN expo-router를 사용하면, Stack을 사용해서 depth를 가지는 navigation이 가능하다*/}
-            {/* Link를 사용하면 push가 가능하다 */}
-            <AuthCheck redirectPath="/pages/Auth/Entrance">
-              <Stack screenOptions={{ headerShown: false }}>
-                <Stack.Screen name="pages/EditPage" />
-                <Stack.Screen name="pages/Auth/Entrance" />
-              </Stack>
-            </AuthCheck>
-          </BackgroundColorProvider>
-        </MoodKeywordProvider>
+        <BackgroundColorProvider>
+          <StatusBar style="dark" />
+          {/* RN expo-router를 사용하면, Stack을 사용해서 depth를 가지는 navigation이 가능하다*/}
+          {/* Link를 사용하면 push가 가능하다 */}
+          <AuthCheck redirectPath="/pages/Auth/Entrance">
+            <Stack screenOptions={{ headerShown: false }}>
+              <Stack.Screen name="pages/EditPage" />
+              <Stack.Screen name="pages/Auth/Entrance" />
+            </Stack>
+          </AuthCheck>
+        </BackgroundColorProvider>
         <Toast position="bottom" config={toastConfig} />
       </ThemeProvider>
     </>
